backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
babel==2.18.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
courlan==1.3.2
cryptography==46.0.6
dateparser==1.4.0
distro==1.9.0
feedparser==6.0.12
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.69.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.4
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
regex==2026.3.32
requests==2.33.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tenacity==9.1.4
tld==0.13.2
tqdm==4.67.3
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
fastapi
uvicorn
